- separate chat room in chat screen
